requirements.txt
python-vlc
yt-dlp
Pillow
